+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-05-29
+
 ### Added
 - Read problems also from standard input.
 - Add feature `mapgen`, which allows map generation with binary `generate-maze`.
@@ -14,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Rename main binary executable to `solve-maze`
 - Maze `hero_start`, `dragon_start` and `goal` are now usize indexes.
+- Errors are separated for cases: no viable paths to end, or no paths without meeting the dragon.
+- If there is no viable path from the dragon position to the hero position, the problem can be solved but the dragon does not move.
 
 ## [0.1.2] - 2024-05-28
 
@@ -37,7 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dragon routing based on shortest path to current hero position.
 - Playback solution.
 
-[unreleased]: https://github.com/mkouhia/wundernut-vol13/compare/v0.1.2...HEAD
+[unreleased]: https://github.com/mkouhia/wundernut-vol13/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/mkouhia/wundernut-vol13/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/mkouhia/wundernut-vol13/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/mkouhia/wundernut-vol13/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/mkouhia/wundernut-vol13/tree/v0.1.0